refactor($core): move core resuable utilities to 'shared-utils and fix tests
