--- packages/@vuepress/core/__test__/prepare/fixtures/docs-custom-theme/README.md
@@ -1,1 +1,0 @@
-# Simple Docs
