--- packages/@vuepress/core/__test__/prepare/fixtures/docs-simple-config/README.md
@@ -1,1 +1,0 @@
-# Simple Docs
