+++ packages/@vuepress/core/lib/webpack/markdownLoader.js
@@ -3,7 +3,8 @@ const path = require('path')
 const hash = require('hash-sum')
 const { EventEmitter } = require('events')
 const { getOptions } = require('loader-utils')
-const { inferTitle, extractHeaders, parseFrontmatter } = require('../util/index')
+const { inferTitle, extractHeaders } = require('../util/index')
+const { parseFrontmatter } = require('@vuepress/shared-utils')
 const LRU = require('lru-cache')
 
 const cache = LRU({ max: 1000 })
