+++ packages/@vuepress/shared-utils/__test__/unescapeHtml.spec.js
@@ -1,0 +1,7 @@
+import unescapeHtml from '../lib/unescapeHtml'
+
+test('should unescape html', () => {
+  const input = '&lt;div&gt;'
+  expect(unescapeHtml(input)).toBe('<div>')
+})
+
