+++ packages/@vuepress/shared-utils/index.js
@@ -1,6 +1,19 @@
 exports.logger = require('./lib/logger')
 exports.env = require('./lib/env')
 exports.codegen = require('./lib/codegen')
+exports.compose = require('./lib/compose')
+exports.datatypes = require('./lib/datatypes')
+exports.parseFrontmatter = require('./lib/parseFrontmatter')
+exports.unescapeHtml = require('./lib/unescapeHtml')
+exports.parseEmojis = require('./lib/parseEmojis')
+exports.parseHeaders = require('./lib/parseHeaders')
+exports.deeplyParseHeaders = require('./lib/deeplyParseHeaders')
+exports.fileToPath = require('./lib/fileToPath')
+exports.sort = require('./lib/sort')
+exports.ensureLeadingSlash = require('./lib/ensureLeadingSlash')
+exports.ensureEndingSlash = require('./lib/ensureEndingSlash')
+exports.getPermalink = require('./lib/getPermalink')
+
 exports.chalk = require('chalk')
 exports.fs = require('fs-extra')
 exports.globby = require('globby')
