+++ packages/@vuepress/shared-utils/lib/compose.js
@@ -1,0 +1,10 @@
+/**
+ * Build functional pipeline.
+ */
+module.exports = function compose (...processors) {
+  if (processors.length === 0) return input => input
+  if (processors.length === 1) return processors[0]
+  return processors.reduce((prev, next) => {
+    return (...args) => next(prev(...args))
+  })
+}
