+++ packages/@vuepress/shared-utils/lib/ensureEndingSlash.js
@@ -1,0 +1,5 @@
+module.exports = function ensureEndingSlash (path) {
+  return /(\.html|\/)$/.test(path)
+    ? path
+    : path + '/'
+}
