+++ packages/@vuepress/shared-utils/lib/ensureLeadingSlash.js
@@ -1,0 +1,3 @@
+module.exports = function ensureLeadingSlash (path) {
+  return path.replace(/^\/?/, '/')
+}
