+++ packages/@vuepress/shared-utils/lib/getPermalink.js
@@ -1,29 +1,5 @@
-const indexRE = /(^|.*\/)(index|readme)\.md$/i
-const extRE = /\.(vue|md)$/
-
-exports.fileToPath = function (file) {
-  if (exports.isIndexFile(file)) {
-    // README.md -> /
-    // foo/README.md -> /foo/
-    return file.replace(indexRE, '/$1')
-  } else {
-    // foo.md -> /foo.html
-    // foo/bar.md -> /foo/bar.html
-    return `/${file.replace(extRE, '').replace(/\\/g, '/')}.html`
-  }
-}
-
-exports.isIndexFile = function (file) {
-  return indexRE.test(file)
-}
-
-exports.sort = function (arr) {
-  return arr.sort((a, b) => {
-    if (a < b) return -1
-    if (a > b) return 1
-    return 0
-  })
-}
+const ensureEndingSlash = require('./ensureEndingSlash')
+const ensureLeadingSlash = require('./ensureLeadingSlash')
 
 // e.g.
 // filename: docs/_posts/evan you.md
@@ -34,12 +10,12 @@ exports.sort = function (arr) {
 // :year/:month/:day/:title/ => 2018/08/14/yyx 990803/
 // :year/:month/:day/:original/ => 2018/08/14/_posts/evan you.html
 
-exports.getPermalink = function ({
+module.exports = function getPermalink ({
   pattern,
   slug,
   date,
   regularPath,
-  localePath
+  localePath = '/'
 }) {
   if (!pattern) {
     return
@@ -71,17 +47,5 @@ exports.getPermalink = function ({
       .replace(/:slug/, slug)
       .replace(/:regular/, regularPath)
 
-  return ensureLeadingSlash(
-    ensureEndingSlash(link)
-  )
-}
-
-function ensureLeadingSlash (path) {
-  return path.replace(/^\/?/, '/')
-}
-
-function ensureEndingSlash (path) {
-  return /(\.html|\/)$/.test(path)
-    ? path
-    : path + '/'
+  return ensureLeadingSlash(ensureEndingSlash(link))
 }
