+++ packages/@vuepress/shared-utils/lib/isIndexFile.js
@@ -1,0 +1,8 @@
+const indexRE = /(^|.*\/)(index|readme)\.md$/i
+
+function isIndexFile (file) {
+  return indexRE.test(file)
+}
+
+module.exports = isIndexFile
+module.exports.indexRE = indexRE
