+++ packages/@vuepress/shared-utils/lib/parseEmojis.js
@@ -1,0 +1,4 @@
+module.exports = str => {
+  const emojiData = require('markdown-it-emoji/lib/data/full.json')
+  return String(str).replace(/:(.+?):/g, (placeholder, key) => emojiData[key] || placeholder)
+}
