+++ packages/@vuepress/shared-utils/lib/parseFrontmatter.js
@@ -1,0 +1,12 @@
+module.exports = function parseFrontmatter (content) {
+  const matter = require('gray-matter')
+  const toml = require('toml')
+
+  return matter(content, {
+    excerpt_separator: '<!-- more -->',
+    engines: {
+      toml: toml.parse.bind(toml),
+      excerpt: false
+    }
+  })
+}
