+++ packages/@vuepress/shared-utils/lib/sort.js
@@ -1,0 +1,7 @@
+module.exports = function sort (arr) {
+  return arr.sort((a, b) => {
+    if (a < b) return -1
+    if (a > b) return 1
+    return 0
+  })
+}
