+++ packages/@vuepress/shared-utils/lib/unescapeHtml.js
@@ -1,0 +1,7 @@
+module.exports = html => String(html)
+  .replace(/&quot;/g, '"')
+  .replace(/&#39;/g, '\'')
+  .replace(/&#x3A;/g, ':')
+  .replace(/&lt;/g, '<')
+  .replace(/&gt;/g, '>')
+
