+++ packages/@vuepress/test-utils/fixtures/docs-local-theme/.vuepress/config.js
@@ -1,4 +1,4 @@
 module.exports = {
   title: 'Hello VuePress',
-  description: 'Just playing around'
+  description: '# Hello, VuePress!'
 }
