+++ packages/@vuepress/test-utils/fixtures/docs-local-theme/README.md
@@ -1,0 +1,1 @@
+# Hello, VuePress!
