+++ packages/@vuepress/test-utils/fixtures/docs-simple-config/README.md
@@ -1,0 +1,1 @@
+# Hello, VuePress!
